doc-auth-service/requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.49
botocore==1.31.49
click==8.1.7
colorama==0.4.6
docutils==0.20.1
docx2txt==0.8
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
psycopg2==2.9.7
mammoth==1.6.0
